fix: drawing structured addresses
According to the [Swiss QR-bill specification][1], a structured
address (type S) has `line1` representing the _street name_ and `line2`
representing the _street number_. When a structured address is provided,
the library currently prints `line1` and `line2` on two separate
lines. However, for type S addresses, `line1` and `line2` should
be printed combined on a single line.

While it is possible to provide the _street name_ and _number_ combined
in `line1` as a workaround, most e-banking software expects these to
be separate. Otherwise, the software may not populate its form
correctly, causing unnecessary friction in the payment process.

[1]: https://www.six-group.com/dam/download/banking-services/standardization/qr-bill/ig-qr-bill-v2.2-en.pdf